3270 - Add CountrySummary getOneOrNone
