+++ src/server/repository/assessmentCycle/countrySummary/index.ts
@@ -1,17 +1,15 @@
 import { createMaterializedView } from './createMaterializedView'
 import { dropMaterializedView } from './dropMaterializedView'
 import { getCount } from './getCount'
-import { getLastEditOdpData } from './getLastEditOdpData'
-import { getLastUpdate } from './getLastUpdate'
 import { getMany } from './getMany'
+import { getOneOrNone } from './getOneOrNone'
 import { refreshMaterializedView } from './refreshMaterializedView'
 
 export const CountrySummaryRepository = {
   createMaterializedView,
   dropMaterializedView,
   getCount,
-  getLastEditOdpData,
-  getLastUpdate,
   getMany,
+  getOneOrNone,
   refreshMaterializedView,
 }
